requirements.txt
llama-cpp-python==0.2.23
jinja2==3.1.2
pylatex
